Add active stock factory

+++ src/main/kotlin/com/dream365/kotlin/model/ActiveStock.kt
@@ -2,13 +2,13 @@ package com.dream365.kotlin.model
 
 data class Change(
     val positive: Boolean,
-    val value: Long
+    val value: Float
 ) {
     companion object {
         fun getChangeFromString(str: String): Change {
             val isPositive = str.contains('+')
             val valueStr: String = if(isPositive) str.replace("+", "") else str.replace("-", "")
-            val value: Long = valueStr.replace("%", "").trim().toLong()
+            val value: Float = valueStr.replace("%", "").trim().toFloat()
 
             return Change(isPositive, value)
         }
@@ -18,7 +18,7 @@ data class Change(
 data class ActiveStock(
     val symbol: String,
     val name: String,
-    val price: Long,
+    val price: Float,
     val change: Change,
     val percentChange: Change
 )
